% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfFromGR.features.R
\name{cfFromGR.features}
\alias{cfFromGR.features}
\title{Calculate the confusion matrix from counts of features shared between two GRanges objects.}
\usage{
cfFromGR.features(query, subject, minPropOverlap = 0.5, verbose = FALSE)
}
\arguments{
\item{query}{a foreground set of regions (perhaps predictions) \code{"GRanges"}}

\item{subject}{a background or reference set of regions \code{"GRanges"}}

\item{minPropOverlap}{a value between 0.0 and 1.0 specifying the proportion of each feature (or gap) in subject covered by query to be counted.}

\item{verbose}{set to TRUE for more output}
}
\value{
\code{"matrix"}
}
\description{
Counts the number of features (TP) or gaps (FP) in a subject (reference) GRanges object
that are covered by the features defined in the query.  Returns a confusion matrix 
listing TP, FP, FN, TN.
}
\examples{
library(GenomicRanges)
# create a mock Seqinfo object with chromosome names and lengths based on SacCer3
test_Si <- x <- Seqinfo(seqnames=c("chrI", "chrII"),
                        seqlengths=c(230218, 813184),
                        isCircular=c(FALSE, FALSE),
                        genome="test_genome")



# create a set of perfectly regular regions. 
regEnd_I <- seq(from=1000, to = seqlengths(test_Si)[1], by=1000 )
regEnd_II <- seq(from=1000, to = seqlengths(test_Si)[2], by=1000 )
gr_reg_I <- GRanges(seqnames = "chrI", ranges = IRanges(end=regEnd_I, width=500), seqinfo=test_Si)
gr_reg_II <- GRanges(seqnames = "chrII", ranges = IRanges(end=regEnd_II, width=500), seqinfo=test_Si)
# need to specify size of chromosomes using seqinfo, borrow this from BSgenome.Scerevisiae.UCSC.sacCer3
(gr_reg <- c(gr_reg_I, gr_reg_II))
width(gr_reg)
gr_regGaps <- gaps(gr_reg, ignore.strand=T )

# runs some checks
countOverlaps(gr_reg, gr_regGaps)  # should all be zero.  Perfect opposites. 
sum(countOverlaps(gr_reg, gr_regGaps) )


gr_fullChroms <- GRanges(seqnames=seqnames(test_Si), ranges=IRanges(start=rep(1, length(seqnames(test_Si))), end=seqlengths(test_Si)), seqinfo=test_Si)
# a bad prediction might predict ALL the chromosomes is in a given state (or none of it). 

# now create a GR with exactly half the chromosome in covered in one single range. 
gr_halfChrom <- GRanges(seqnames=seqnames(test_Si), ranges=IRanges(start=rep(1, length(seqnames(test_Si))), end=floor(seqlengths(test_Si)/2)), seqinfo=test_Si)


cfFromGR.features(gr_halfChrom, gr_reg)

cfFromGR.features(gr_reg, gr_reg)  # perfect overlap

cfFromGR.features(gr_halfChrom, gr_halfChrom)  ### perfect overlap between two very large regions. 

cfFromGR.features(gr_fullChroms, gr_reg)   # zero negative predictions 
cfFromGR.features(gr_regGaps, gr_reg)    # perfect missing

cfFromGR.features(gr_halfChrom, gr_regGaps) 
cfFromGR.features(gr_regGaps, gr_halfChrom ) 
cfFromGR.features(gr_regGaps, gr_halfChrom , minPropOverlap = 0.001)   # will call pretty much any overlap as a positive. 



}
\seealso{
\code{\link{cfFromGR}}
}
