% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfFromGR.R
\name{cfFromGR}
\alias{cfFromGR}
\title{Calculate the confusion matrix from two GRanges objects.}
\usage{
cfFromGR(
  query,
  subject,
  method = c("features", "bases"),
  minPropOverlap = 0.5,
  maxgap = -1L,
  minoverlap = 0L,
  genomeSize = 0L,
  verbose = FALSE
)
}
\arguments{
\item{query}{a foreground set of regions (perhaps predictions) \code{"GRanges"}}

\item{subject}{a background or reference set of regions \code{"GRanges"}}

\item{method}{whether to count "features" [default] or all "bases"}

\item{minPropOverlap}{a value between 0.0 and 1.0 specifying the proportion of each feature (or gap) in subject covered by query to be counted.}

\item{maxgap}{how much to extend the query by to count an overlap.}

\item{minoverlap}{Number of bases of overlap to count in method="bases"}

\item{genomeSize}{0L used with method="bases" to provide total genomeSize}

\item{verbose}{set to TRUE for more output}
}
\value{
\code{"matrix"}
}
\description{
Returns a confusion matrix 
listing TP, FP, FN, TN.
}
\examples{
library(GenomicRanges)
# create a mock Seqinfo object with chromosome names and lengths based on SacCer3
test_Si <- x <- Seqinfo(seqnames=c("chrI", "chrII"),
                        seqlengths=c(230218, 813184),
                        isCircular=c(FALSE, FALSE),
                        genome="test_genome")


# create a set of perfectly regular GenomicRanges regions. 
regEnd_I <- seq(from=1000, to = seqlengths(test_Si)[1], by=1000 )
regEnd_II <- seq(from=1000, to = seqlengths(test_Si)[2], by=1000 )
gr_reg_I <- GRanges(seqnames = "chrI", ranges = IRanges(end=regEnd_I, width=500), seqinfo=test_Si)
gr_reg_II <- GRanges(seqnames = "chrII", ranges = IRanges(end=regEnd_II, width=500), seqinfo=test_Si)
# need to specify size of chromosomes using seqinfo
(gr_reg <- c(gr_reg_I, gr_reg_II))
width(gr_reg)
gr_regGaps <- gaps(gr_reg, ignore.strand=T )

# runs some checks
countOverlaps(gr_reg, gr_regGaps)  # should all be zero.  Perfect opposites. 
sum(countOverlaps(gr_reg, gr_regGaps) )

gr_fullChroms <- GRanges(seqnames=seqnames(test_Si), ranges=IRanges(start=rep(1, length(seqnames(test_Si))), end=seqlengths(test_Si)), seqinfo=test_Si)
# a bad prediction might predict ALL the chromosomes is in a given state (or none of it). 

# now create a GR with exactly half the chromosome in covered in one single range. 
gr_halfChrom <- GRanges(seqnames=seqnames(test_Si), ranges=IRanges(start=rep(1, length(seqnames(test_Si))), end=floor(seqlengths(test_Si)/2)), seqinfo=test_Si)



cfFromGR(query = gr_halfChrom, subject = gr_reg) # approx equal in each class
cfFromGR(query = gr_halfChrom, subject = gr_reg, method="bases", ) 


cfFromGR(gr_reg, gr_reg, verbose=T) # perfect overlap
cfFromGR(gr_reg, gr_reg, verbose=T)


cfFromGR(gr_halfChrom, gr_halfChrom, verbose=T)  ### perfect overlap between two very large regions. 
cfFromGR(gr_halfChrom, gr_halfChrom, verbose=T)

# zero negative predictions 
cfFromGR(gr_fullChroms, gr_reg)
cfFromGR(gr_fullChroms, gr_reg, method="bases", verbose=TRUE)

# perfect missing
cfFromGR(gr_regGaps, gr_reg) 
cfFromGR(gr_regGaps, gr_reg, method="bases", verbose=TRUE) 

# subject is a very large domain. 
cfFromGR(gr_regGaps, gr_halfChrom )  # misses one because overlap < minPropOverlap
# will call pretty much any overlap as a positive. 
cfFromGR(gr_regGaps, gr_halfChrom , minPropOverlap = 0.001) 


}
