% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBindingFactor.DNA_motif.R
\name{createBindingFactor.DNA_motif}
\alias{createBindingFactor.DNA_motif}
\title{Create a binding factor object to match a given DNA motif}
\usage{
createBindingFactor.DNA_motif(
  name,
  type = "DNA_motif",
  patternString = "N",
  patternLength = nchar(patternString),
  stateWidth = patternLength,
  profile.layers = "LAYER.1",
  profile.marks = 0,
  mod.layers = "LAYER.1",
  mod.marks = 1,
  test.layer0.binding = FALSE,
  test.mismatch.rate = 0.1,
  max.pattern.tries = 1000,
  min.DM.length = 2,
  min.DR.length = 10,
  verbose = FALSE,
  max.mismatch = 0,
  min.mismatch = 0,
  with.indels = FALSE,
  fixed = TRUE,
  algorithm = "auto"
)
}
\arguments{
\item{name}{give the binding factor a name}

\item{type}{["DNA_motif"]  to differentiate from other types}

\item{patternString}{["N"] put motif here (using IUPAC codes for degenerate bases)}

\item{patternLength}{length of pattern to be matched [nchar(patternString)]}

\item{stateWidth}{the width of pattern to recognise on other layers}

\item{profile.layers}{a vector of named layers to set as a match}

\item{profile.marks}{a vector of 0/1 to match the layers in profile.layers}

\item{mod.layers}{a vector of named layers to alter on a match}

\item{mod.marks}{a vector of 0/1 to set on the mod.layers}

\item{test.layer0.binding}{when creating, test if the DNA sequence has a match.}

\item{test.mismatch.rate}{proportion of mismatches to tolerate when testing [.1]}

\item{max.pattern.tries}{NA}

\item{min.DM.length}{NA}

\item{verbose}{set to TRUE for more output}

\item{max.mismatch}{0   see \code{\link{vmatchPattern}}}

\item{min.mismatch}{0   see \code{\link{vmatchPattern}}}

\item{with.indels}{FALSE  see \code{\link{vmatchPattern}}}

\item{fixed}{TRUE  see \code{\link{vmatchPattern}}}

\item{algorithm}{"auto"  see \code{\link{vmatchPattern}}}

\item{min.DR.lengt}{NA}
}
\value{
\code{"bindingFactor"}
}
\description{
Create a new binding factor based on a DNA motif that \emph{may}  also require 
marks on others layers and \emph{may} (when used) set marks on other layers.
Makes use of biostrings function \code{\link{vmatchPattern}}
Can use IUPAC codes and allow mismatches
}
\examples{
DNA_A <- createBindingFactor.DNA_motif(name="DNA_A",patternString = "CAT" )

gc1-2 <- createBindingFactor.DNA_motif(name="gc1-2",
                          patternString="SSSSSSSSSSSSSSSS",
                          max.mismatch= 2, 
                          min.mismatch= 1,
                          fixed="subject",
                          profile.layers = NULL,
                          profile.marks=NULL,
                          mod.layers="LAYER.1", mod.marks = 1)

}
\seealso{
\code{\link{runLayerBinding}} \code{\link{createBindingFactor.DNA_regexp}}
}
