% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLayerList.DNAString.R
\name{createLayerList.DNAstring}
\alias{createLayerList.DNAstring}
\title{Create a LayerList object from a DNA sequence}
\usage{
createLayerList.DNAstring(seq, layerNames = NULL,
  nLayers = length(layerNames), verbose = TRUE)
}
\arguments{
\item{seq}{method to do something to (\code{"hsv"} or \code{"cluster"})}

\item{layerNames}{description of that param}

\item{nLayers}{=10 you get the idea}

\item{verbose}{you get the idea}
}
\value{
\code{"LayerList"}
}
\description{
Create a LayerList object from a DNAString object
}
\examples{
x <- 1   # great!

}

