% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLayerSet.BSgenome.R
\name{createLayerSet.BSgenome}
\alias{createLayerSet.BSgenome}
\title{Create a genome-wide layerList object}
\usage{
createLayerSet.BSgenome(genome, n.layers = 1, layer.names = paste("LAYER",
  1:n.layers, sep = "."), verbose = FALSE)
}
\arguments{
\item{genome}{a BSgenome objects}

\item{layer.names=paste("LAYER", 1:n.layers, }{sep=".")}

\item{verbose=FALSE}{}
}
\value{
\code{"LayerList"}
}
\description{
Convenience wrapper to quickly create a layerList from a BSgenome object.
}
\examples{
x <- 1   # great!

}
\seealso{
\code{\link{runLayerBinding.BSgenome}}
}
