% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRandomBindingFactor.R
\name{createRandomBindingFactor}
\alias{createRandomBindingFactor}
\title{Randomly generate a binding factor}
\usage{
createRandomBindingFactor(
  name,
  layerSet,
  type = c("DNA_motif", "DNA_region", "layer_region", "layer_island"),
  test.layer0.binding = FALSE,
  test.mismatch.rate = 0.1,
  max.pattern.tries = 1000,
  min.DM.length = 2,
  min.DR.length = 10,
  verbose = FALSE
)
}
\arguments{
\item{name}{give the binding factor a name}

\item{type}{[c("DNA_motif", "DNA_region","layer_region","layer_island")]  to differentiate from other types}

\item{test.layer0.binding}{when creating, test if the DNA sequence has a match.}

\item{test.mismatch.rate}{proportion of mismatches to tolerate when testing [.1]}

\item{max.pattern.tries}{1000}

\item{min.DM.length}{2}

\item{verbose}{set to TRUE for more output}

\item{patternString}{["N"] put motif here (using IUPAC codes for degenerate bases)}

\item{patternLength}{length of pattern to be matched [nchar(patternString)]}

\item{stateWidth}{the width of pattern to recognise on other layers}

\item{profile.layers}{a vector of named layers to set as a match}

\item{profile.marks}{a vector of 0/1 to match the layers in profile.layers}

\item{mod.layers}{a vector of named layers to alter on a match}

\item{mod.marks}{a vector of 0/1 to set on the mod.layers}

\item{min.DR.lengt}{10}
}
\value{
\code{"bindingFactor"}
}
\description{
Create a new binding factor that \emph{may}  also require 
marks on others layers and \emph{may} (when used) set marks on other layers.
Need to specify the type of factor.
}
\examples{
rBf <- createRandomBindingFactor(name="DNA_A",type="DNA_motif",patternString = "CAT" )

}
\seealso{
\code{\link{runLayerBinding}} \code{\link{createBindingFactor.DNA_regexp}}
}
