% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHitsCache.R
\name{generateHitsCache}
\alias{generateHitsCache}
\title{Generate a cache of binding factor within a LayerList}
\usage{
generateHitsCache(layerList, factorSet, cache.layers = NULL, verbose = FALSE)
}
\arguments{
\item{layerList}{a \code{"Layerlist"} object containing a layerSet based on a BSgenome and other meta-data}

\item{factorSet}{a \code{"list"} of \code{"bindingFactor"} objects}

\item{cache.layers}{which named layers to cache hits on (default NULL)}

\item{verbose}{give more output}
}
\value{
\code{"LayerList"}
}
\description{
Generate and attach a cache of binding factor hit positions for named binding factors on certain levels.
This is an internal function called from runLayerBinding.BSgenome 
In turn it uses \code{\link{matchBindingFactor.BSgenome}} to generate hits.
}
\examples{
require(Biostrings)
require(BSgenome.Scerevisiae.UCSC.sacCer3)

genome <- BSgenome.Scerevisiae.UCSC.sacCer3   # for convenience


scLayerSet <- createLayerSet.BSgenome(genome=genome, n.layers = 5, verbose=TRUE)

testFactor3 <- createBindingFactor.DNA_motif("test", patternString="ACTGGGCTA", profile.layers = c("LAYER.1", "LAYER.3"), profile.marks = c(0,0), 
                                             mod.layers = c("LAYER.2", "LAYER.4"), mod.marks=c(0,1))

generateHitsCache(layerList=scLayerSet, factorSet=list(testFactor3))

}
\seealso{
\code{\link{runLayerBinding.BSgenome}} \code{\link{matchBindingFactor.BSgenome}}
}
