% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSM1182890_d3.xist.mus.meanScores}
\alias{GSM1182890_d3.xist.mus.meanScores}
\title{Mean scores from mm9 chrX}
\format{A vector of Mean scores from mm9 chrX}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3904790/}
}
\usage{
GSM1182890_d3.xist.mus.meanScores
}
\description{
Mean scores from mm9 chrX
GSM1182890_d3.xist.mus.meanScores
}
\keyword{datasets}

