% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepBSgenomeSequences.R
\name{keepBSgenomeSequences}
\alias{keepBSgenomeSequences}
\title{Subset a BSgenome object}
\usage{
keepBSgenomeSequences(genome, seqnames)
}
\arguments{
\item{genome}{a BSgenome object}

\item{seqnames}{names of chromosomes to keep}
}
\value{
\code{"genome"}
}
\description{
Create a BSgenome object with some sequences (chromosomes) removed. 
This is useful when wanting to work without mitochondria or unplaced chromosomes.
This function was posted by Hervé Pagès on the bioconductor support pages around 2016
\link{https://support.bioconductor.org/p/83588/}
}
\examples{

library(BSgenome.Scerevisiae.UCSC.sacCer3)
genome <- BSgenome.Scerevisiae.UCSC.sacCer3
seqinfo(genome)
seqnames(genome)
sequences_to_keep <- setdiff(seqnames(genome), "chrM")   # nuclear chroms only
genomeNuc <- keepBSgenomeSequences(genome, sequences_to_keep)
seqinfo(genomeNuc)

}
