% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchBindingFactor.BSgenome.R
\name{matchBindingFactor.BSgenome}
\alias{matchBindingFactor.BSgenome}
\title{Find matches for a binding factor on a layer set containing a BSgenome sequence}
\usage{
matchBindingFactor.BSgenome(layerSet, bindingFactor,
  match.layers = names(bindingFactor$profile), clusterGap = 10,
  max.window = 1e+07, cache.layers = NULL, verbose = FALSE)
}
\arguments{
\item{layerSet}{the \code{"layerSet"} target}

\item{bindingFactor}{the \code{"bindingFactor"} to match}

\item{match.layers}{restrict matches to only these named layers (default: all layers in names(bindingFactor$profile))}

\item{clusterGap}{=10 NOT IMPLEMENTED}

\item{max.window}{=10000000 on less powerful computers, break up the search into windows of this size.}

\item{verbose}{output more information to the screen}
}
\value{
\code{"hits"}
}
\description{
Generate a list of matches for a binding factor against a layerSet object containing a BSgenome sequence.
}
\examples{
x <- 1   # great!

}
