% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchRegExp.BSgenome.R
\name{matchRegExp.BSgenome}
\alias{matchRegExp.BSgenome}
\title{Customised matching for regular expression binding factors to BSgenomes}
\usage{
matchRegExp.BSgenome(genome, forRegExp, revRegExp)
}
\arguments{
\item{genome}{a \code{"BSgenome"} object}

\item{forRegExp}{see \code{\link{createBindingFactor.DNA_regexp}}}

\item{revRegExp}{see \code{\link{createBindingFactor.DNA_regexp}}}
}
\value{
\code{"Hits"}
}
\description{
Generate a list of matches for a binding factor against a BSgenome object. 
With proper forward and reverse patterns, will return hits from both strands.
}
\seealso{
\code{\link{createBindingFactor.DNA_regexp}}
}
