% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyLayerByBindingFactor.Views.R
\name{modifyLayerByBindingFactor.Views}
\alias{modifyLayerByBindingFactor.Views}
\title{Modify a layerSet object}
\usage{
modifyLayerByBindingFactor.Views(layerSet, hits, bindingFactor,
  verbose = FALSE)
}
\arguments{
\item{layerSet}{method to do something to (\code{"hsv"} or \code{"cluster"})}

\item{hits}{a \code{"Views"} object representing the places the bindingFactor can bind.}

\item{bindingFactor}{The \code{"bindingFactor"}}

\item{verbose}{you get the idea}
}
\value{
\code{"layerSet"}
}
\description{
Modify a layerSet object according to a bindingFactor and a pre-calculated set of hits.
}
\examples{
x <- 1   # great!

}

