% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.factorSet.R
\name{plot.factorSet}
\alias{plot.factorSet}
\title{Plot to visualise a set of binding factor profiles and mod layers}
\usage{
\method{plot}{factorSet}(factorSet)
}
\arguments{
\item{factorSet}{the list of binding factors.}
}
\description{
show the profile and modification specificities for a set of binding factors.
}
\examples{
testFactor2 <- createBindingFactor.DNA_motif("test", patternString="ACTGGGCTA")

testFactor3 <- createBindingFactor.DNA_motif("test", patternString="ACTGGGCTA", profile.layers = c("LAYER.1", "LAYER.3"), profile.marks = c(0,0), 
                                             mod.layers = c("LAYER.2", "LAYER.4"), mod.marks=c(0,1))

# check that a profile looking for 1 will not find any.  N.B this WILL bind AFTER testFactor2
testFactor4 <- createBindingFactor.DNA_motif("test", patternString="ACTGGGCTA", profile.layers = c("LAYER.3", "LAYER.4"), profile.marks = c(0,1), 
                                             mod.layers = c("LAYER.1", "LAYER.2"), mod.marks=c(0,1))

testFactor5 <- createBindingFactor.layer_region("test5", patternLength = 150)
# now can match things genome wide. Need to run layerBinding and modification.

# need to have a factorSet, a list of bindingFactors

testFS <- list(testFactor2=testFactor2, testFactor3=testFactor3, testFactor4=testFactor4, testFactor5=testFactor5)

plot.factorSet(testFS)

}
\seealso{
\code{\link{createBindingFactor.DNA_motif}} \code{\link{print.bfSet}}
}
