% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.layerBindingHistory.R
\name{plot.layerBindingHistory}
\alias{plot.layerBindingHistory}
\title{Plot the accumulation of coverage and blocks after a simulation}
\usage{
\method{plot}{layerBindingHistory}(modLayerSet)
}
\arguments{
\item{modLayerSet}{A layerset produced by runLayerBinding()}
}
\value{
a plot
}
\description{
Plot the accumulation of coverage and blocks after a simulation
}
