% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.score.hits.R
\name{plot.score.hits}
\alias{plot.score.hits}
\title{[unfinsihed] plot multiple score measures. Green for good, red for bad.}
\usage{
\method{plot}{score.hits}(
  query,
  target,
  methods = c("acc", "tpr", "tnr", "ppv", "npv", "fpr", "fnr"),
  colours = structure(c("green", "green", "green", "green", "green", "red", "red"), names
    = c("acc", "tpr", "tnr", "ppv", "npv", "fpr", "fnr"))
)
}
\arguments{
\item{query}{}

\item{target}{}

\item{methods}{}

\item{colours}{}
}
\value{
a plot
}
\description{
[unfinsihed] plot multiple score measures. Green for good, red for bad.
}
