% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bfSet.R
\name{print.bfSet}
\alias{print.bfSet}
\title{Formatted printing for a set of binding factors}
\usage{
\method{print}{bfSet}(factorSet)
}
\arguments{
\item{factorSet}{method to do something to (\code{"hsv"} or \code{"cluster"})}
}
\description{
Formatted printing for a set of binding factors
}
\examples{
x <- 1   # great!

}

