% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randGrangesBigGenome.R
\name{randGrangesBigGenome}
\alias{randGrangesBigGenome}
\title{randGrangesBigGenome}
\usage{
randGrangesBigGenome(
  genome,
  n = NULL,
  sizeFunc = NULL,
  gapFunc = NULL,
  argsSizeFunc,
  argsGapFunc,
  minGapSize = 1,
  minRange = 1,
  verbose = FALSE
)
}
\arguments{
\item{genome}{A BSgenome object}

\item{n}{NULL}

\item{sizeFunc}{NULL  must be a function that generates a vector of positive integers}

\item{gapFunc}{NULL must be a function that generates a vector of positive integers}

\item{argsSizeFunc}{NULL a list giving named parameter values passed to sizeFunc}

\item{argsGapFunc}{NULL a list giving named parameter values passed to gapFunc}

\item{minGapSize}{1}

\item{minRange}{1}

\item{verbose}{FALSE  Give more output}
}
\value{
\code{"GRanges"} object with seqinfo matching genome
}
\description{
Generate random non-overlapping GRanges features within parameters of size and distance
}
\examples{

library(BSgenome.Scerevisiae.UCSC.sacCer3)

genome <- BSgenome.Scerevisiae.UCSC.sacCer3




test <- randGrangesBigGenome(genome=genome, 
                             sizeFunc = function(value, n) rep(x=value, times=n), 
                             gapFunc = function(value, n) rep(x=value, times=n), 
                             argsSizeFunc = list(value=147, n=5),
                             argsGapFunc = list(value=30, n=7), verbose=T)
# fixed nuclesomes of 147bp with random gaps following a poisson.
testGR <- randGrangesBigGenome(genome=genome, 
                               sizeFunc = function(value, n) rep(x=value, times=n), 
                               gapFunc = function(n, value) rpois(n = n , lambda = value), 
                               argsSizeFunc = list(value=147, n=5),
                               argsGapFunc = list(value=40, n=7))




}
