% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLayerBinding.BSgenome.R
\name{runLayerBinding.BSgenome}
\alias{runLayerBinding.BSgenome}
\title{Run the layer binding simulation on a BSgenome object}
\usage{
runLayerBinding.BSgenome(layerList, factorSet, iterations = 1,
  bindingFactorFreqs = rep(1, length(factorSet)),
  watch.function = function(x) { }, collect.stats = FALSE,
  target.layer = 2, verbose = FALSE, ...)
}
\arguments{
\item{layerList}{a \code{"Layerlist"} object containing a layerSet and other meta-data}

\item{factorSet}{a \code{"list"} of \code{"bindingFactor"} objects}

\item{iterations}{how many changes to make  NEED TO RENAME THIS}

\item{bindingFactorFreqs}{the relative proportions of each \code{"bindingFactor"} in \code{"factorSet"}}

\item{watch.function}{have this function execute during each iteration e.g. print something}

\item{collect.stats}{collect a table of stats each iteration}

\item{target.layer}{NOT IMPLEMENTED}

\item{verbose}{give more output}
}
\value{
\code{"LayerList"}
}
\description{
Simulate layer binding by the binding factor on the layerList
}
\examples{
x <- 1   # great!

}
\seealso{
\code{\link{runLayerBinding}}
}
