% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBindingFactor.DNA_motif.R
\name{createBindingFactor.DNA_motif}
\alias{createBindingFactor.DNA_motif}
\title{Create a binding factor object to match a given DNA motif}
\usage{
createBindingFactor.DNA_motif(name, type = "DNA_motif", patternString = "N",
  patternLength = nchar(patternString), stateWidth = patternLength,
  profile.layers = "LAYER.1", profile.marks = 0, mod.layers = "LAYER.1",
  mod.marks = 1, test.layer0.binding = FALSE, test.mismatch.rate = 0.1,
  max.pattern.tries = 1000, min.DM.length = 2, min.DR.length = 10,
  verbose = FALSE)
}
\arguments{
\item{name}{give the binding factor a name}

\item{type}{= "DNA_motif"  to differentiate from other types}

\item{patternString}{= "N" put motif here (using IUPAC codes for degenerate bases)}

\item{patternLength}{= nchar(patternString)}

\item{stateWidth}{the width of pattern to recognise on other layers}

\item{profile.layers}{a vector of named layers to set as a match}

\item{profile.marks}{a vector of 0/1 to match the layers in profile.layers}

\item{mod.layers}{a vector of named layers to alter on a match}

\item{mod.marks}{a vector of 0/1 to set on the mod.layers}

\item{test.layer0.binding}{when creating, test if the DNA sequence has a match.}

\item{test.mismatch.rate}{% mismatches to tolerate when testing}

\item{max.pattern.tries}{NA}

\item{min.DM.length}{NA}

\item{verbose}{set to TRUE for more output}

\item{min.DR.lengt}{NA}
}
\value{
\code{"bindingFactor"}
}
\description{
Create a new binding factor based on a DNA motif that \emph{may}  also require 
marks on others layers and \emph{may} (when used) set marks on other layers.
}
\examples{
DNA_A <- createBindingFactor.DNA_motif(name="DNA_A",patternString = "CAT" )

}
\seealso{
\code{\link{runLayerBinding}} \code{\link{createBindingFactor.DNA_regexp}}
}
