% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBindingFactor.DNA_motif.R
\name{createBindingFactor.DNA_motif}
\alias{createBindingFactor.DNA_motif}
\title{Find matches for a binding factor on a layer set}
\usage{
createBindingFactor.DNA_motif(name, type = "DNA_motif", patternString = "N",
  patternLength = nchar(patternString), stateWidth = patternLength,
  profile.layers = "LAYER.1", profile.marks = 0, mod.layers = "LAYER.1",
  mod.marks = 1, test.layer0.binding = FALSE, test.mismatch.rate = 0.1,
  max.pattern.tries = 1000, min.DM.length = 2, min.DR.length = 10,
  verbose = FALSE)
}
\arguments{
\item{name}{method to do something to (\code{"hsv"} or \code{"cluster"})}

\item{type}{description of that param}

\item{patternString}{=10 you get the idea}

\item{patternLength}{=10000000 you get the idea}

\item{stateWidth}{you get the idea}

\item{profile.layers}{you get the idea}

\item{profile.marks}{you get the idea}

\item{mod.layers}{you get the idea}

\item{mod.marks}{you get the idea}

\item{test.layer0.binding}{you get the idea}

\item{test.mismatch.rate}{you get the idea}

\item{max.pattern.tries}{you get the idea}

\item{min.DM.length}{you get the idea}

\item{verbose}{you get the idea}

\item{min.DR.lengt}{you get the idea}
}
\value{
\code{"hits"}
}
\description{
Generate a list of matches for a binding factor against a layerSet object.
}
\examples{
x <- 1   # great!

}

