% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBindingFactor.layer_region.R
\name{createBindingFactor.layer_region}
\alias{createBindingFactor.layer_region}
\title{Create a binding factor object to match a given pattern of layers, typically ignoring DNA}
\usage{
createBindingFactor.layer_region(
  name,
  type = "layer_region",
  patternLength = 1,
  patternString = "N",
  mismatch.rate = 0,
  stateWidth = patternLength,
  profile.layers = "LAYER.1",
  profile.marks = 0,
  mod.layers = "LAYER.1",
  mod.marks = 1,
  offset = 0,
  offset.method = NULL,
  align = "centre",
  test.layer0.binding = FALSE,
  test.mismatch.rate = 0.1,
  max.pattern.tries = 1000,
  min.DM.length = 2,
  min.DR.length = 10,
  verbose = FALSE
)
}
\arguments{
\item{name}{give the binding factor a name}

\item{type}{"layer_region"  to differentiate from other types}

\item{patternLength}{[= 1] length of pattern}

\item{patternString}{= NOT USED in this case}

\item{stateWidth}{the width of pattern to recognise on other layers, default is same as patternLength}

\item{profile.layers}{a vector of named layers to set as a match}

\item{profile.marks}{a vector of 0/1 to match the layers in profile.layers}

\item{mod.layers}{a vector of named layers to alter on a match}

\item{mod.marks}{a vector of 0/1 to set on the mod.layers}

\item{test.layer0.binding}{when creating, test if the DNA sequence has a match.}

\item{test.mismatch.rate}{proportion of mismatches to tolerate when testing}

\item{max.pattern.tries}{NA}

\item{min.DM.length}{NA}

\item{min.DR.length}{NA}

\item{verbose}{set to TRUE for more output}
}
\value{
\code{"hits"}
}
\description{
Create a new binding factor based on a simple pattern of 
marks on others layers and \emph{may} (when used) set marks on other layers.
}
\examples{
bf.LR <- createBindingFactor.layer_region("layerBf", type="layer_region", 
         patternLength = 1, patternString = "N",
          stateWidth = patternLength, profile.layers = "LAYER.1",
          profile.marks = 0, mod.layers = "LAYER.1", mod.marks = 1)

}
