% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBindingFactor.layer_region.R
\name{createBindingFactor.layer_region}
\alias{createBindingFactor.layer_region}
\title{Create a new biding factor}
\usage{
createBindingFactor.layer_region(name, type = "layer_region",
  patternLength = 1, patternString = "N", mismatch.rate = 0,
  stateWidth = patternLength, profile.layers = "LAYER.1",
  profile.marks = 0, mod.layers = "LAYER.1", mod.marks = 1, offset = 0,
  offset.method = NULL, align = "centre", test.layer0.binding = FALSE,
  test.mismatch.rate = 0.1, max.pattern.tries = 1000, min.DM.length = 2,
  min.DR.length = 10, verbose = FALSE)
}
\arguments{
\item{name}{method to do something to (\code{"hsv"} or \code{"cluster"})}

\item{type}{description of that param}

\item{patternLength}{=10000000 you get the idea}

\item{patternString}{=10 you get the idea}

\item{stateWidth}{you get the idea}

\item{profile.layers}{you get the idea}

\item{profile.marks}{you get the idea}

\item{mod.layers}{you get the idea}

\item{mod.marks}{you get the idea}

\item{test.layer0.binding}{you get the idea}

\item{test.mismatch.rate}{you get the idea}

\item{max.pattern.tries}{you get the idea}

\item{min.DM.length}{you get the idea}

\item{verbose}{you get the idea}

\item{min.DR.lengt}{you get the idea}
}
\value{
\code{"hits"}
}
\description{
Create a new biding factor
}
\examples{
x <- 1   # great!

}

