% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLayerList.DNAString.R
\name{createLayerList.DNAstring}
\alias{createLayerList.DNAstring}
\title{Create a LayerList object from a DNA sequence}
\usage{
createLayerList.DNAstring(
  seq,
  layerNames = NULL,
  nLayers = length(layerNames),
  verbose = TRUE
)
}
\arguments{
\item{seq}{a (\code{"DNAString"} object}

\item{layerNames}{a character vector of names for the layers. Should be unique}

\item{nLayers}{How many layers to make. Can be specified _instead_ of _layerNames_}

\item{verbose}{Output extra information}
}
\value{
\code{"LayerList"}
}
\description{
Create a LayerList object from a DNAString object
}
\examples{
library(Biostrings)
mySeq <- DNAString("TGACATCGTCTATCGATCG")
createLayerList.DNAstring(seq=mySeq, nLayers=1)

}
