% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLayerSet.BSgenome.R
\name{createLayerSet.BSgenome}
\alias{createLayerSet.BSgenome}
\title{Create a genome-wide layerList object}
\usage{
createLayerSet.BSgenome(
  genome,
  n.layers = 1,
  layer.names = paste("LAYER", 1:n.layers, sep = "."),
  verbose = FALSE
)
}
\arguments{
\item{genome}{a BSgenome objects}

\item{n.layers}{1}

\item{layer.names}{paste("LAYER",1:n.layers, sep=".")}

\item{verbose}{Give more output [FALSE]}
}
\value{
\code{"LayerList"}
}
\description{
Convenience wrapper to quickly create a layerList from a BSgenome object.
}
\examples{
require(Biostrings)
require(BSgenome.Scerevisiae.UCSC.sacCer3)  # a relatively small genome

genome <- BSgenome.Scerevisiae.UCSC.sacCer3 
scLayerSet <- createLayerSet.BSgenome(genome, n.layers=5)

scLayerSet2 <- createLayerSet.BSgenome(genome, layer.names=c("H3K4me1", "H3K27me3"))

}
\seealso{
\code{\link{runLayerBinding.BSgenome}}
}
