% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHitsCache.R
\name{generateHitsCache}
\alias{generateHitsCache}
\title{Generate a cache of binding factor within a LayerList}
\usage{
generateHitsCache(layerList, factorSet, cache.layers = NULL, verbose = FALSE)
}
\arguments{
\item{layerList}{a \code{"Layerlist"} object containing a layerSet and other meta-data}

\item{factorSet}{a \code{"list"} of \code{"bindingFactor"} objects}

\item{cache.layers}{which named layers to cache hits on (default NULL)}

\item{verbose}{give more output}
}
\value{
\code{"LayerList"}
}
\description{
Generate and attach a cache of binding factor hit positions for named binding factors on certain levels.
}
\examples{
x <- 1   # great!

}
\seealso{
\code{\link{runLayerBinding}}
}
