% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchBindingFactor.R
\name{matchBindingFactor}
\alias{matchBindingFactor}
\title{Find matches for a binding factor on a layer set}
\usage{
matchBindingFactor(layerSet, bindingFactor, clusterGap = 10,
  max.window = 1e+07, verbose = FALSE)
}
\arguments{
\item{layerSet}{method to do something to (\code{"hsv"} or \code{"cluster"})}

\item{bindingFactor}{description of that param}

\item{clusterGap}{=10 you get the idea}

\item{max.window}{=10000000 you get the idea}

\item{verbose}{you get the idea}
}
\value{
\code{"hits"}
}
\description{
Generate a list of matches for a binding factor against a layerSet object.
}
\examples{
x <- 1   # great!

}

