% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLayerBinding.BSgenome.R
\name{runLayerBinding.BSgenome}
\alias{runLayerBinding.BSgenome}
\title{Run the layer binding simulation on a BSgenome object}
\usage{
runLayerBinding.BSgenome(layerList, factorSet, iterations = 1,
  bindingFactorFreqs = rep(1, length(factorSet)),
  watch.function = function(x) { }, collect.stats = FALSE,
  target.layer = 2, verbose = FALSE, ...)
}
\arguments{
\item{layerList}{method to do something to (\code{"hsv"} or \code{"cluster"})}

\item{factorSet}{description of that param}

\item{iterations}{you get the idea}

\item{bindingFactorFreqs}{you get the idea}

\item{watch.function}{you get the idea}

\item{collect.stats}{you get the idea}

\item{target.layer}{you get the idea}

\item{verbose}{you get the idea}
}
\value{
\code{"LayerList"}
}
\description{
Description text here
}
\examples{
x <- 1   # great!

}

